Convert pkg/config/utils into module
