+++ pkg/config/utils/endpoints_test.go
@@ -3,9 +3,10 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package setup
+package utils
 
 import (
+	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"testing"
 
 	"github.com/DataDog/datadog-agent/comp/core/secrets"
@@ -19,10 +20,10 @@ import (
 // endpoints are configured.
 // Refer to https://github.com/DataDog/viper/pull/2 for more details.
 func TestSecretBackendWithMultipleEndpoints(t *testing.T) {
-	conf := Conf()
+	conf := pkgconfigsetup.Conf()
 	conf.SetConfigFile("./tests/datadog_secrets.yaml")
 	// load the configuration
-	_, err := LoadDatadogCustom(conf, "datadog_secrets.yaml", optional.NewNoneOption[secrets.Component](), nil)
+	_, err := pkgconfigsetup.LoadDatadogCustom(conf, "datadog_secrets.yaml", optional.NewNoneOption[secrets.Component](), nil)
 	assert.NoError(t, err)
 
 	expectedKeysPerDomain := map[string][]string{
@@ -45,7 +46,7 @@ additional_endpoints:
   - someapikey
 `
 
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -77,7 +78,7 @@ additional_endpoints:
   - someapikey
 `
 
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -100,7 +101,7 @@ func TestGetMultipleEndpointsEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_ADDITIONAL_ENDPOINTS", "{\"https://foo.datadoghq.com\": [\"someapikey\"]}")
 
-	testConfig := Conf()
+	testConfig := pkgconfigsetup.Conf()
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -130,7 +131,7 @@ additional_endpoints:
   - someapikey
 `
 
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -157,7 +158,7 @@ dd_url: "https://app.datadoghq.com"
 api_key: fakeapikey
 `
 
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -185,7 +186,7 @@ additional_endpoints:
   - ""
 `
 
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -218,7 +219,7 @@ additional_endpoints:
   - someapikey
 `
 
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 
@@ -240,7 +241,7 @@ additional_endpoints:
 func TestSiteEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_SITE", "datadoghq.eu")
-	testConfig := ConfFromYAML("")
+	testConfig := pkgconfigsetup.ConfFromYAML("")
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
@@ -260,7 +261,7 @@ func TestDefaultSite(t *testing.T) {
 	datadogYaml := `
 api_key: fakeapikey
 `
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
@@ -281,7 +282,7 @@ func TestSite(t *testing.T) {
 site: datadoghq.eu
 api_key: fakeapikey
 `
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
@@ -301,7 +302,7 @@ func TestDDURLEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_URL", "https://app.datadoghq.eu")
 	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := ConfFromYAML("")
+	testConfig := pkgconfigsetup.ConfFromYAML("")
 	testConfig.BindEnv("external_config.external_agent_dd_url")
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
@@ -322,7 +323,7 @@ func TestDDDDURLEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_DD_URL", "https://app.datadoghq.eu")
 	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := ConfFromYAML("")
+	testConfig := pkgconfigsetup.ConfFromYAML("")
 	testConfig.BindEnv("external_config.external_agent_dd_url")
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
@@ -347,7 +348,7 @@ func TestDDURLAndDDDDURLEnvVar(t *testing.T) {
 	t.Setenv("DD_URL", "https://app.datadoghq.dd_url.eu")
 
 	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := ConfFromYAML("")
+	testConfig := pkgconfigsetup.ConfFromYAML("")
 	testConfig.BindEnv("external_config.external_agent_dd_url")
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
@@ -373,7 +374,7 @@ api_key: fakeapikey
 external_config:
   external_agent_dd_url: "https://external-agent.datadoghq.com"
 `
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
@@ -397,7 +398,7 @@ api_key: fakeapikey
 external_config:
   external_agent_dd_url: "https://custom.external-agent.datadoghq.eu"
 `
-	testConfig := ConfFromYAML(datadogYaml)
+	testConfig := pkgconfigsetup.ConfFromYAML(datadogYaml)
 
 	multipleEndpoints, err := GetMultipleEndpoints(testConfig)
 	externalAgentURL := GetMainEndpoint(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
