+++ pkg/config/utils/miscellaneous.go
@@ -9,11 +9,11 @@ package utils
 import (
 	"path/filepath"
 
-	"github.com/DataDog/datadog-agent/pkg/config"
+	pkgconfigmodel "github.com/DataDog/datadog-agent/pkg/config/model"
 )
 
 // ConfFileDirectory returns the absolute path to the folder containing the config
 // file used to populate the registry
-func ConfFileDirectory(c config.Reader) string {
+func ConfFileDirectory(c pkgconfigmodel.Reader) string {
 	return filepath.Dir(c.ConfigFileUsed())
 }
