+++ pkg/config/utils/telemetry.go
@@ -5,19 +5,21 @@
 
 package utils
 
-import "github.com/DataDog/datadog-agent/pkg/config"
+import (
+	pkgconfigmodel "github.com/DataDog/datadog-agent/pkg/config/model"
+)
 
 // IsCheckTelemetryEnabled returns if we want telemetry for the given check.
 // Returns true if a * is present in the telemetry.checks list.
-func IsCheckTelemetryEnabled(checkName string) bool {
+func IsCheckTelemetryEnabled(checkName string, cfg pkgconfigmodel.Reader) bool {
 	// false if telemetry is disabled
-	if !IsTelemetryEnabled() {
+	if !IsTelemetryEnabled(cfg) {
 		return false
 	}
 
 	// by default, we don't enable telemetry for every checks stats
-	if config.Datadog.IsSet("telemetry.checks") {
-		for _, check := range config.Datadog.GetStringSlice("telemetry.checks") {
+	if cfg.IsSet("telemetry.checks") {
+		for _, check := range cfg.GetStringSlice("telemetry.checks") {
 			if check == "*" {
 				return true
 			} else if check == checkName {
@@ -29,6 +31,6 @@ func IsCheckTelemetryEnabled(checkName string) bool {
 }
 
 // IsTelemetryEnabled returns whether or not telemetry is enabled
-func IsTelemetryEnabled() bool {
-	return config.Datadog.IsSet("telemetry.enabled") && config.Datadog.GetBool("telemetry.enabled")
+func IsTelemetryEnabled(cfg pkgconfigmodel.Reader) bool {
+	return cfg.IsSet("telemetry.enabled") && cfg.GetBool("telemetry.enabled")
 }
