+++ tasks/modules.py
@@ -172,6 +172,7 @@ DEFAULT_MODULES = {
     "pkg/config/model": GoModule("pkg/config/model", independent=True),
     "pkg/config/env": GoModule("pkg/config/env", independent=True),
     "pkg/config/setup": GoModule("pkg/config/setup", independent=True),
+    "pkg/config/utils": GoModule("pkg/config/utils", independent=True),
     "pkg/config/logs": GoModule("pkg/config/logs", independent=True),
     "pkg/config/remote": GoModule("pkg/config/remote", independent=True),
     "pkg/security/secl": GoModule("pkg/security/secl", independent=True),
